TweetSearchApi/requirements.txt
feedparser==5.2.1
requests-oauthlib==1.3.0
Flask==1.1.2
pytz==2020.1
mysql-connector-python==8.0.20